requirements.txt
beautifulsoup4==4.9.3
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
google-api-core==1.22.4
google-auth==1.22.1
google-cloud-core==1.4.2
google-cloud-firestore==1.2.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Janome==0.4.1
Jinja2==2.11.2
joblib==1.0.0
MarkupSafe==1.1.1
nltk==3.5
oauthlib==3.1.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PySocks==1.7.1
python-dotenv==0.15.0
pytz==2020.1
regex==2020.11.13
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
textblob==0.15.3
tokenizer==2.4.0
tqdm==4.56.0
tweepy==3.10.0
urllib3==1.25.10
Werkzeug==1.0.1
